refactor(projects): extract calculateProjectTotals into dedicated service module
Move business logic out of db.ts into src/lib/services/project.service.ts
and update all imports across routes, tests, and mocks.
